requirements.txt
asgiref==3.4.1
Django==3.2.7
django-betterforms==1.2
django-debug-toolbar==3.2.2
django-dotenv==1.4.2
Pillow==8.3.2
psycopg2==2.9.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
